requirements.txt
gspread
google-auth
openai
apscheduler
twilio
pydantic
python-dotenv
requests
pytz
pandas